+++ packages/heroku-pipelines/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- Add `pipelines:transfer` command to set and update the pipeline owner.
+
 ## [2.3.1] 2017-09-13
 
 - Fixes serialization for `pipelines:info --json`. https://github.com/heroku/heroku-pipelines/pull/72 changed the serialization for that command changing `apps` to `pipelineApps`.